# postman/openApiSpecification.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Playmatch API",
        "description": "Playmatch API documentation",
        "termsOfService": "",
        "contact": {
            "name": "",
            "url": "",
            "email": ""
        },
        "license": {
            "name": "unlicensed",
            "url": ""
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:9000",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/tipo-documentos/{id}": {
            "get": {
                "tags": [
                    "tipo-documento-resource"
                ],
                "operationId": "getTipoDocumento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoDocumentoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "tipo-documento-resource"
                ],
                "operationId": "updateTipoDocumento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoDocumentoDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoDocumentoDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "tipo-documento-resource"
                ],
                "operationId": "deleteTipoDocumento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "patch": {
                "tags": [
                    "tipo-documento-resource"
                ],
                "operationId": "partialUpdateTipoDocumento",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoDocumentoDTO"
                            }
                        },
                        "application/merge-patch+json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoDocumentoDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoDocumentoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tipo-documentos": {
            "get": {
                "tags": [
                    "tipo-documento-resource"
                ],
                "operationId": "getAllTipoDocumentos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Zero-based page index (0..N)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "description": "The size of the page to be returned",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TipoDocumentoDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "tipo-documento-resource"
                ],
                "operationId": "createTipoDocumento",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TipoDocumentoDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/TipoDocumentoDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TipoDocumentoDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "sigla": {
                        "type": "string",
                        "maxLength": 20,
                        "minLength": 0
                    },
                    "nombreDocumento": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 0
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "ACTIVO",
                            "INACTIVO"
                        ]
                    }
                },
                "required": [
                    "estado",
                    "nombreDocumento",
                    "sigla"
                ]
            },
...
        }
    }
}